requirements.txt
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.0
google-cloud-dialogflow==2.17.0
vk-api==11.9.9